+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Pending
+
+* Make it possible to focus on a  hierarchy / tree of tests using `focused` on context level.
+
 ## 0.1.0 (2019-10-17)
 
 * Initial public release